+++ src/SemigroupK.ts
@@ -14,7 +14,7 @@
  *
  * @since 3.0.0
  */
-import type { Lazy } from './function'
+import type { LazyArg } from './function'
 import type { HKT, Kind, Typeclass } from './HKT'
 
 // -------------------------------------------------------------------------------------
@@ -27,7 +27,7 @@ import type { HKT, Kind, Typeclass } from './HKT'
  */
 export interface SemigroupK<F extends HKT> extends Typeclass<F> {
   readonly combineK: <S, R2, W2, E2, B>(
-    second: Lazy<Kind<F, S, R2, W2, E2, B>>
+    second: LazyArg<Kind<F, S, R2, W2, E2, B>>
   ) => <R1, W1, E1, A>(self: Kind<F, S, R1, W1, E1, A>) => Kind<F, S, R1 & R2, W1 | W2, E1 | E2, A | B>
 }